requirements.txt
django==3.2.13
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
drf-extensions==0.7.1
